add retest method to htsohm main file
